재출시 투표 기능 구현

+++ backend_set/src/main/java/org/funding/retryVotes/service/RetryVotesService.java
@@ -1,6 +1,8 @@
 package org.funding.retryVotes.service;
 
 import lombok.RequiredArgsConstructor;
+import org.funding.fund.dao.FundDAO;
+import org.funding.fund.vo.FundVO;
 import org.funding.retryVotes.dao.RetryVotesDAO;
 import org.funding.retryVotes.dto.DoVoteRequestDTO;
 import org.funding.retryVotes.vo.RetryVotesVO;
@@ -11,6 +13,7 @@ import org.springframework.stereotype.Service;
 public class RetryVotesService {
 
     private final RetryVotesDAO retryVotesDAO;
+    private final FundDAO fundDAO;
 
     // 투표 등록
     public String doVote(DoVoteRequestDTO voteRequestDTO) {
@@ -25,12 +28,32 @@ public class RetryVotesService {
         retryVotesVO.setFundingId(voteRequestDTO.getFundId());
         retryVotesDAO.addRetryVotes(retryVotesVO);
 
+        // 펀딩에 투표 수 추가
+        FundVO fund = fundDAO.selectById(voteRequestDTO.getFundId());
+        int count = fund.getRetryVotesCount();
+        if (count > 0) {
+            fund.setRetryVotesCount(fund.getRetryVotesCount() + 1);
+        }
+
         return "투표가 정상적으로 등록되었습니다.";
     }
 
     // 투표 취소
     public String deleteVote(DoVoteRequestDTO voteRequestDTO) {
+        boolean alreadyVoted = retryVotesDAO.existsVoteByUserIdAndFundingId(voteRequestDTO.getUserId(), voteRequestDTO.getFundId());
+        // 투표 취소방지
+        if (!alreadyVoted) {
+            throw new IllegalStateException("투표를 안하셨기에 투표를 취소 할 수 없습니다.");
+        }
+
         retryVotesDAO.deleteRetryVotes(voteRequestDTO);
+
+        // 펀딩에 투표 수 취소 반영
+        FundVO fund = fundDAO.selectById(voteRequestDTO.getFundId());
+        int count = fund.getRetryVotesCount();
+        if (count > 0) {
+            fund.setRetryVotesCount(fund.getRetryVotesCount() - 1);
+        }
         return "투표가 정상적으로 취소되었습니다";
     }
